requirements.txt
attrs==20.3.0
aws-cdk.assets==1.95.1
aws-cdk.aws-apigateway==1.95.1
aws-cdk.aws-autoscaling-common==1.95.1
aws-cdk.aws-cloudformation==1.95.1
aws-cdk.aws-cloudwatch==1.95.1
aws-cdk.aws-cognito==1.95.1
aws-cdk.aws-dynamodb==1.95.1
aws-cdk.aws-ec2==1.95.1
aws-cdk.aws-ecr==1.95.1
aws-cdk.aws-ecr-assets==1.95.1
aws-cdk.aws-efs==1.95.1
aws-cdk.aws-events==1.95.1
aws-cdk.aws-iam==1.95.1
aws-cdk.aws-kms==1.95.1
aws-cdk.aws-lambda==1.95.1
aws-cdk.aws-logs==1.95.1
aws-cdk.aws-route53==1.95.1
aws-cdk.aws-s3==1.95.1
aws-cdk.aws-s3-assets==1.95.1
aws-cdk.aws-signer==1.95.1
aws-cdk.aws-sns==1.95.1
aws-cdk.aws-sqs==1.95.1
aws-cdk.aws-ssm==1.95.1
aws-cdk.core==1.95.1
aws-cdk.custom-resources==1.95.1
aws-cdk.cx-api==1.95.1
aws-cdk.region-info==1.95.1
cattrs==1.3.0
constructs==3.3.71
jsii==1.26.0
publication==0.0.3
python-dateutil==2.8.1
six==1.15.0
typing-extensions==3.7.4.3
